Start moving to @Rule-based test suite
This allows for tests to avoid having explicit iterations, and
allows for tests to pass iff they pass a given percentage of the
iterations. This seems to clear up most problems relating to numerical
imprecision, because a single error caused by imprecision out of a
thousand iterations does not fail the entire build.